build(deps): update ramenctl to v0.3.4
Fixes includes in this version:
- https://github.com/RamenDR/ramenctl/issues/56
- https://github.com/RamenDR/ramenctl/pull/96
- https://github.com/RamenDR/ramenctl/pull/93

ramenctl requires now go 1.23.5, fixing CVE-2024-45336 (because of
github.com/ramendr/ramen/e2e). `go mod tidy` updated the toolchain to
go1.24.1, but I change it to go1.23.7 to minimize the chance of picking
up packages that are not compatible with the downstream build.

With this version applications deployed by `odf dr test run` command can
be viewed and manipulated in OpenShift console. This can be useful if
the `odf dr test clean` was not run.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>