cmd: introduce the devpreview package
The devpreview package helps to configure commands as developer preview;
unsupported and not fully tested. We want to make sure that users are
aware about such commands.

The semantics of developer preview are explained here:
https://access.redhat.com/support/offerings/devpreview

The package provides constants and functions to make it easy to
configure commands as developer preview and get consistent behavior in
multiple commands.

The configure a command and all its children as developer preview call:

    devpreview.Configure(cmd)

Command configured as developer preview has the following behavior:

The command is mark hidden, so we don't display it in online help of the
parent command. The user can still use the command if they learn about
the name of the command from documentation or from support:

    odf command-name --help

Sub commands are not hidden, to keep online help useful if you know the
name of the command.

The short description of the command includes a `(developer preview)`
suffix. This is seen in help text of the parent command:

    root    root short (developer preview)

The long description of the command includes a developer preview note.
This is see in the command help text:

    root long description

    +------------------------------------------------------------------------------+
    |                                                                              |
    |    This command is a developer preview, unsupported and not fully tested.    |
    |    Please see the following document for more info on developer preview:     |
    |    https://access.redhat.com/support/offerings/devpreview                    |
    |                                                                              |
    +------------------------------------------------------------------------------+

We will use this package for the dr command introduced in 4.19 as
developer preview.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>