dr: mark the dr commands as developer preview
The dr command is hidden so we don't display it in in odf --help. To use
the command the user will need to learn about it in documentation or
instruction from support.

The dr commands include a clear note about the status of the command:

    $ odf dr --help
    Troubleshoot ODF DR (developer preview)

    +------------------------------------------------------------------------------+
    |                                                                              |
    |    This command is a developer preview, unsupported and not fully tested.    |
    |    Please see the following document for more info on developer preview:     |
    |    https://access.redhat.com/support/offerings/devpreview                    |
    |                                                                              |
    +------------------------------------------------------------------------------+

    Usage:
      odf dr [command]

    ...

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit 29de6ab53759b8231c45ca2d54d1010cb35725bf)